Further breakdown into discrete python mixin modules: apt and chroot
