+++ pyfll/builder.py
@@ -7,29 +7,27 @@ import glob
 import hashlib
 import logging
 import os
-import shlex
 import shutil
 import subprocess
 import sys
-import tarfile
 import tempfile
 import threading
 import time
 
 from configobj import ConfigObj, flatten_errors
-from debian.debian_support import Version
 from validate import Validator
 
+from pyfll.apt import AptMixin
 from pyfll.bootloader import BootloaderMixin
-from pyfll.exceptions import FllError, FllLocalesError
+from pyfll.chroot import ChrootExecMixin
+from pyfll.exceptions import FllError
 from pyfll.gpt import run_gpthybrid
 from pyfll.isodd import write_iso
-from pyfll.locales import FllLocales
-from pyfll.profile import FllProfile, parse_dependency_groups
-from pyfll.util import deduplicate_list, multiline_to_list, uuidgen
+from pyfll.profile import PackageProfileMixin
+from pyfll.util import uuidgen
 
 
-class FLLBuilder(BootloaderMixin):
+class FLLBuilder(BootloaderMixin, AptMixin, PackageProfileMixin, ChrootExecMixin):
     env = {
         "LANGUAGE": "C",
         "LC_ALL": "C",
@@ -206,258 +204,6 @@ class FLLBuilder(BootloaderMixin):
         self.conf.write()
         os.chown(self.conf.filename, self.opts.uid, self.opts.gid)
 
-    def expand_pkg_profile(
-        self, chroot: str, profile: str, modules_dir: str
-    ) -> FllProfile:
-        """Return a FllProfile for a given chroot and profile."""
-        pkg_profile = FllProfile()
-        for package in self.conf["chroots"][chroot]["packages"].get("packages"):
-            pkg_profile.packages.add(package)
-        arch = self.conf["chroots"][chroot]["packages"]["arch"]
-        linux = self.conf["chroots"][chroot]["packages"]["linux"]
-        browsers = self.conf["chroots"][chroot]["packages"]["browser"]
-        for browser in browsers:
-            pkg_profile.packages.add(browser)
-
-        ro_fs = self.conf["options"]["readonly_filesystem"]
-        if ro_fs == "squashfs":
-            pkg_profile.packages.add("squashfs-tools")
-        elif ro_fs == "erofs":
-            pkg_profile.packages.add("erofs-utils")
-
-        initramfs_tool = self.conf["options"]["initramfs_tool"]
-        pkg_profile.packages.add(initramfs_tool)
-
-        bootloader = self.conf["options"]["bootloader"]
-        if bootloader == "grub" or bootloader == "grub-efi":
-            if arch in ("amd64", "i386") and bootloader == "grub":
-                pkg_profile.packages.add("grub-pc")
-            if arch == "amd64":
-                pkg_profile.packages.add("grub-efi-amd64-bin")
-            elif arch == "i386":
-                pkg_profile.packages.add("grub-efi-ia32-bin")
-            elif arch == "arm64":
-                pkg_profile.packages.add("grub-efi-arm64-bin")
-        elif bootloader == "systemd-boot":
-            pkg_profile.packages.update(["systemd-boot", "systemd-boot-efi"])
-        elif bootloader == "refind":
-            pkg_profile.packages.add("refind")
-            pkg_profile.debconf.add("refind refind/install_to_esp boolean false")
-
-        linux_meta = ["linux-image", "linux-headers"]
-        pkg_profile.packages.update(
-            ["-".join([prefix, linux]) for prefix in linux_meta]
-        )
-
-        pkg_profile.flatpaks = set(
-            self.conf["chroots"][chroot]["flatpak"]["flathub"]["flatpaks"]
-        )
-        pkg_profile.flatpaks_beta = set(
-            self.conf["chroots"][chroot]["flatpak"]["flathub-beta"]["flatpaks"]
-        )
-
-        fll_profile_spec = os.path.join(self.opts.share, "fll.profile.spec")
-        profile_conf = ConfigObj(profile, configspec=fll_profile_spec)
-        self.validate_configobj(profile_conf)
-
-        if "desc" in profile_conf:
-            for line in multiline_to_list(profile_conf["desc"]):
-                self.log.debug(f"  {line}")
-
-        if "debconf" in profile_conf:
-            self.log.debug("debconf:")
-            for line in multiline_to_list(profile_conf["debconf"]):
-                pkg_profile.debconf.add(line)
-                self.log.debug(f"  {line}")
-
-        if "packages" in profile_conf:
-            self.log.debug("packages:")
-            for line in multiline_to_list(profile_conf["packages"]):
-                pkg_profile.packages.add(line)
-                self.log.debug(f"  {line}")
-
-        packages_arch = f"packages_{arch}"
-        if packages_arch in profile_conf:
-            self.log.debug(f"packages_{arch}:")
-            for line in multiline_to_list(profile_conf[packages_arch]):
-                pkg_profile.packages.add(line)
-                self.log.debug(f"  {line}")
-
-        if "flatpaks" in profile_conf:
-            self.log.debug("flatpaks:")
-            for line in multiline_to_list(profile_conf["flatpaks"]):
-                pkg_profile.flatpaks.add(line)
-                self.log.debug(f"  {line}")
-
-        if "flatpaks_beta" in profile_conf:
-            self.log.debug("flatpaks_beta:")
-            for line in multiline_to_list(profile_conf["flatpaks_beta"]):
-                pkg_profile.flatpaks_beta.add(line)
-                self.log.debug(f"  {line}")
-
-        if "desktops" in profile_conf:
-            self.log.debug("desktops:")
-            for line in multiline_to_list(profile_conf["desktops"]):
-                pkg_profile.desktops.add(line)
-                self.log.debug(f"  {line}")
-
-        if "groups" in profile_conf:
-            self.log.debug("groups:")
-            for line in multiline_to_list(profile_conf["groups"]):
-                pkg_profile.groups.add(line)
-                self.log.debug(f"  {line}")
-
-        modules = set()
-        if "modules" in profile_conf:
-            self.log.debug("modules:")
-            for module in multiline_to_list(profile_conf["modules"]):
-                modules.add(module)
-                self.log.debug(f"  {module}")
-
-        if "modules" in self.conf["chroots"][chroot]["packages"]:
-            self.log.debug("modules (config):")
-            for module in self.conf["chroots"][chroot]["packages"]["modules"]:
-                modules.add(module)
-                self.log.debug(f"  {module}")
-
-        if os.path.isfile(profile + ".postinst"):
-            self.log.debug(f"registering postinst script: {profile}.postinst")
-            pkg_profile.postinst.add(profile + ".postinst")
-
-        self.log.debug("---")
-        fll_module_spec = os.path.join(self.opts.share, "fll.module.spec")
-        for module in modules:
-            module_file = os.path.join(modules_dir, module)
-
-            if not os.path.isfile(module_file):
-                self.log.critical(f"no such module file: {module_file}")
-                raise FllError
-
-            module_conf = ConfigObj(module_file, configspec=fll_module_spec)
-            self.validate_configobj(module_conf)
-
-            if "desc" in module_conf:
-                for line in multiline_to_list(module_conf["desc"]):
-                    self.log.debug(f"  {line}")
-
-            if "debconf" in module_conf:
-                self.log.debug("debconf:")
-                for line in multiline_to_list(module_conf["debconf"]):
-                    pkg_profile.debconf.add(line)
-                    self.log.debug(f"  {line}")
-
-            if "packages" in module_conf:
-                self.log.debug("packages:")
-                for line in multiline_to_list(module_conf["packages"]):
-                    pkg_profile.packages.add(line)
-                    self.log.debug(f"  {line}")
-
-            packages_arch = f"packages_{arch}"
-            if packages_arch in module_conf:
-                self.log.debug(f"packages_{arch}:")
-                for line in multiline_to_list(module_conf[packages_arch]):
-                    pkg_profile.packages.add(line)
-                    self.log.debug(f"  {line}")
-
-            if "flatpaks" in module_conf:
-                self.log.debug("flatpaks:")
-                for line in multiline_to_list(module_conf["flatpaks"]):
-                    pkg_profile.flatpaks.add(line)
-                    self.log.debug(f"  {line}")
-
-            if "flatpaks_beta" in module_conf:
-                self.log.debug("flatpaks_beta:")
-                for line in multiline_to_list(module_conf["flatpaks_beta"]):
-                    pkg_profile.flatpaks_beta.add(line)
-                    self.log.debug(f"  {line}")
-
-            if "desktops" in module_conf:
-                self.log.debug("desktops:")
-                for line in multiline_to_list(module_conf["desktops"]):
-                    pkg_profile.desktops.add(line)
-                    self.log.debug(f"  {line}")
-
-            if "groups" in module_conf:
-                self.log.debug("groups:")
-                for line in multiline_to_list(module_conf["groups"]):
-                    pkg_profile.groups.add(line)
-                    self.log.debug(f"  {line}")
-
-            if os.path.isfile(module_file + ".postinst"):
-                self.log.debug(f"registering postinst script: {module_file}.postinst")
-                pkg_profile.postinst.add(module_file + ".postinst")
-
-            self.log.debug("---")
-
-        if any([pkg_profile.flatpaks, pkg_profile.flatpaks_beta]):
-            pkg_profile.packages.add("flatpak")
-
-        return pkg_profile
-
-    def parse_package_profile(self, chroot: str) -> FllProfile:
-        """Parse packages profile for each chroot."""
-        profiles = self.conf["chroots"][chroot]["packages"]["profile"]
-        profile_dir = os.path.join(self.opts.share, "profiles")
-        modules_dir = os.path.join(self.opts.share, "modules")
-
-        chroot_profile = FllProfile()
-        for profile_name in profiles:
-            self.log.info(f"{chroot} - processing package profile: {profile_name}")
-            profile_path = os.path.join(profile_dir, profile_name)
-            if not os.path.isfile(profile_path):
-                self.log.critical(f"no such package profile: {profile_path}")
-                raise FllError
-            chroot_profile.merge(
-                self.expand_pkg_profile(chroot, profile_path, modules_dir)
-            )
-
-        self.log.debug(f"debconf summary for {chroot}:")
-        for item in sorted(chroot_profile.debconf):
-            self.log.debug(f"  {item}")
-
-        self.log.debug(f"desktops summary for {chroot}:")
-        for item in sorted(chroot_profile.desktops):
-            self.log.debug(f"  {item}")
-
-        self.log.debug(f"package summary for {chroot}:")
-        for item in sorted(chroot_profile.packages):
-            self.log.debug(f"  {item}")
-
-        self.log.debug(f"flatpaks summary for {chroot}:")
-        for item in sorted(chroot_profile.flatpaks):
-            self.log.debug(f"  {item}")
-
-        self.log.debug(f"flatpaks_beta summary for {chroot}:")
-        for item in sorted(chroot_profile.flatpaks_beta):
-            self.log.debug(f"  {item}")
-
-        self.log.debug(f"groups summary for {chroot}:")
-        for item in sorted(chroot_profile.groups):
-            self.log.debug(f"  {item}")
-
-        self.log.debug(f"postinst summary for {chroot}:")
-        for item in sorted(chroot_profile.postinst):
-            self.log.debug(f"  {item}")
-
-        return chroot_profile
-
-    def configure_locales(self, chroot: str) -> None:
-        """Generate locales."""
-        chroot_dir = os.path.join(self.temp, chroot)
-        locales_list = (
-            [loc for loc in self.conf["chroots"][chroot]["packages"]["locales"]]
-            if self.conf["chroots"][chroot]["packages"].get("locales")
-            else self.opts.locales
-        )
-        default_locale = f"{locales_list[0]}.UTF-8"
-        with open(os.path.join(chroot_dir, "etc", "locale.gen"), "a") as locale_gen:
-            locale_gen.write("\n# Locales enabled by fll\n")
-            for locale in locales_list:
-                self.log.debug(f"enabling locale: {locale}.UTF-8 UTF-8")
-                locale_gen.write(f"{locale}.UTF-8 UTF-8\n")
-        self.chroot_exec(chroot, ["locale-gen"])
-        self.chroot_exec(chroot, ["update-locale", f"LANG={default_locale}"])
-
     def init_build_directory(self) -> None:
         """Prepare temporary directory for chroots and result staging area."""
         self.log.debug("preparing build area...")
@@ -526,311 +272,6 @@ class FLLBuilder(BootloaderMixin):
 
         if not self.opts.preserve:
             self.nuke_directory(self.temp)
-
-    def exec_cmd(self, cmd: list) -> None:
-        """Execute subprocess, always writing stdout+stderr to the log."""
-        self.log.debug(shlex.join(cmd))
-
-        log_it = self.log.info if self.opts.verbose else self.log.debug
-
-        try:
-            proc = subprocess.Popen(
-                cmd,
-                env=self.env,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT,
-                universal_newlines=True,
-            )
-            for line in iter(proc.stdout.readline, ""):
-                log_it(line.rstrip())
-            proc.stdout.close()
-            return_code = proc.wait()
-            if return_code:
-                raise subprocess.CalledProcessError(return_code, shlex.join(cmd))
-        except KeyboardInterrupt:
-            raise FllError
-        except subprocess.CalledProcessError:
-            self.log.exception(f"problem executing command: {shlex.join(cmd)}")
-            raise FllError
-
-    def chroot_exec(self, chroot: str, args: list) -> None:
-        """Run command in a chroot via systemd-nspawn."""
-        chroot_dir = os.path.join(self.temp, chroot)
-        cmd = [
-            "systemd-nspawn",
-            "--quiet",
-            f"--directory={chroot_dir}",
-            "--as-pid2",
-            "--resolv-conf=bind-host",
-            "--timezone=off",
-            "--restrict-address-families=AF_INET AF_INET6 AF_UNIX",
-        ]
-        for key, value in self.env.items():
-            cmd.append(f"--setenv={key}={value}")
-        cmd.append("--")
-        cmd.extend(args)
-        self.exec_cmd(cmd)
-
-    def chroot_output(self, chroot: str, args: list) -> str:
-        """Run command in a chroot and return captured stdout."""
-        chroot_dir = os.path.join(self.temp, chroot)
-        cmd = [
-            "systemd-nspawn",
-            "--quiet",
-            f"--directory={chroot_dir}",
-            "--as-pid2",
-            "--resolv-conf=bind-host",
-            "--timezone=off",
-            "--restrict-address-families=AF_INET AF_INET6 AF_UNIX",
-        ]
-        for key, value in self.env.items():
-            cmd.append(f"--setenv={key}={value}")
-        cmd.append("--")
-        cmd.extend(args)
-        self.log.debug(shlex.join(cmd))
-        result = subprocess.run(cmd, env=self.env, capture_output=True, text=True)
-        if result.returncode != 0:
-            self.log.critical(result.stderr.strip())
-            raise FllError
-        return result.stdout
-
-    def apt_get(
-        self, chroot: str, command: str, args: list = [], insecure: bool = False
-    ) -> None:
-        """An apt-get install wrapper. Automatic installation of recommended
-        packages defaults to disabled."""
-        aptget = ["apt-get", "--yes"]
-        if insecure:
-            aptget.append("--allow-unauthenticated")
-            aptget.extend(["-o", "Acquire::AllowInsecureRepositories=1"])
-        aptget.extend(["-o", "Acquire::Languages=none"])
-        aptget.extend(["-o", "Dpkg::Use-Pty=0"])
-        aptget.extend(["-o", "Dpkg::Progress-Fancy=0"])
-        aptget.extend(["-o", "APT::Color=0"])
-
-        apt_recommends = self.conf["options"].get("apt_recommends", "no")
-        if apt_recommends == "no":
-            aptget.extend(["-o", "APT::Install-Recommends=0"])
-
-        if self.opts.debug:
-            aptget.extend(["-o", "APT::Get::Show-Versions=1"])
-        if self.opts.verbose:
-            aptget.append("-q")
-        else:
-            aptget.append("-qq")
-
-        aptget.append(command)
-        if args:
-            aptget.extend(args)
-
-        self.chroot_exec(chroot, aptget)
-
-    def debbootstrap(
-        self, chroot: str, arch: str, target: str, mirror: str, codename: str
-    ) -> None:
-        """Bootstrap a distro root filesystem with debootstrap."""
-        bootstrapper = self.conf["options"].get("bootstrapper")
-        bootstrap_includes = "apt-utils,ca-certificates,gnupg,xz-utils,zstd"
-        if bootstrapper == "mmdebstrap":
-            cmd = [
-                "mmdebstrap",
-                f"--architectures={arch}",
-                f"--include={bootstrap_includes}",
-                "--variant=minbase",
-                "--mode=root",
-                "--format=directory",
-                "--hook-dir=/usr/share/mmdebstrap/hooks/merged-usr",
-            ]
-        elif bootstrapper == "debootstrap":
-            cmd = [
-                "debootstrap",
-                f"--arch={arch}",
-                f"--include={bootstrap_includes}",
-                "--variant=minbase",
-                "--merged-usr",
-            ]
-        else:
-            cmd = [
-                "cdebootstrap",
-                f"--arch={arch}",
-                f"--include={bootstrap_includes}",
-                "--flavour=minimal",
-            ]
-
-        cmd.extend([codename, target, mirror])
-
-        if self.opts.debug or self.opts.verbose:
-            cmd.insert(1, "--verbose")
-
-        self.exec_cmd(cmd)
-
-        if bootstrapper == "cdebootstrap":
-            self.chroot_exec(chroot, ["dpkg", "--purge", "cdebootstrap-helper-rc.d"])
-
-    def chroot_bootstrap(self, chroot: str) -> None:
-        """Bootstrap a distro root filesystem with cdebootstrap."""
-        distro = self.conf["chroots"][chroot]["packages"]["distro"]
-        codename = self.conf["chroots"][chroot]["packages"]["codename"]
-        arch = self.conf["chroots"][chroot]["packages"]["arch"]
-
-        dist_repo = self.conf["chroots"][chroot]["repos"][distro]
-        if dist_repo.get("cached"):
-            mirror = dist_repo["cached"]
-        else:
-            mirror = dist_repo["uri"]
-            auto_apt_proxy = shutil.which("auto-apt-proxy")
-            if auto_apt_proxy:
-                apt_proxy = subprocess.run(
-                    [auto_apt_proxy], stdout=subprocess.PIPE
-                ).stdout.decode("utf-8")
-                if apt_proxy:
-                    mirror = apt_proxy.rstrip() + "/" + mirror.split("//")[1]
-
-        target = os.path.join(self.temp, chroot)
-
-        self.log.info(f"{chroot} - bootstrapping {distro} {codename} {arch}...")
-        self.debbootstrap(chroot, arch, target, mirror, codename)
-        self.write_file(chroot, "/etc/hosts")
-        os.mkdir(os.path.join(target, "disks"), 0o755)
-        os.mkdir(os.path.join(target, "fll"), 0o755)
-
-    def write_apt_lists(
-        self, chroot: str, cached: bool = False, src_uri: bool = False
-    ) -> None:
-        """Write apt source lists to /etc/apt/sources.list.d/*."""
-        distro = self.conf["chroots"][chroot]["packages"]["distro"]
-        chroot_dir = os.path.join(self.temp, chroot)
-        auto_apt_proxy = shutil.which("auto-apt-proxy")
-        apt_proxy = None
-        if auto_apt_proxy:
-            apt_proxy = subprocess.run(
-                [auto_apt_proxy], stdout=subprocess.PIPE
-            ).stdout.decode("utf-8")
-        for dist_repo in self.conf["chroots"][chroot]["repos"].keys():
-            repo = self.conf["chroots"][chroot]["repos"][dist_repo]
-            repo_uri = repo.get("uri")
-            cached_uri = repo.get("cached")
-            if not cached_uri and apt_proxy and repo_uri:
-                cached_uri = apt_proxy.rstrip() + "/" + repo_uri.split("//")[1]
-            sources_uri = repo.get("sources_uri")
-            if sources_uri:
-                cmd = ["wget", "--quiet", sources_uri, "-O"]
-                cmd.append(
-                    os.path.join(
-                        chroot_dir,
-                        "etc/apt/sources.list.d",
-                        os.path.basename(sources_uri),
-                    )
-                )
-                self.exec_cmd(cmd)
-
-                if cached and cached_uri:
-                    cmd = ["sed", "-i", f"s#^URIs: .*#URIs: {cached_uri}#"]
-                    cmd.append(
-                        os.path.join(
-                            chroot_dir,
-                            "etc/apt/sources.list.d",
-                            os.path.basename(sources_uri),
-                        )
-                    )
-                    self.exec_cmd(cmd)
-                continue
-
-            sources_file = os.path.join(
-                chroot_dir, "etc/apt/sources.list.d", dist_repo + ".sources"
-            )
-
-            self.log.debug("creating %s" % sources_file)
-            sources_file_fh = None
-            try:
-                sources_file_fh = open(sources_file, "w")
-                if src_uri:
-                    sources_file_fh.write("Types: deb deb-src\n")
-                else:
-                    sources_file_fh.write("Types: deb\n")
-                if cached and cached_uri:
-                    sources_file_fh.write(f"URIs: {cached_uri}\n")
-                else:
-                    sources_file_fh.write(f"URIs: {repo_uri}\n")
-                sources_file_fh.write(f"Suites: {repo['suite']}\n")
-                sources_file_fh.write(f"Components: {repo['components']}\n")
-                if repo.get("keyring"):
-                    sources_file_fh.write(
-                        f"Signed-by: /usr/share/keyrings/{repo['keyring']}.gpg\n"
-                    )
-                else:
-                    sources_file_fh.write(
-                        f"Signed-by: /usr/share/keyrings/{distro}-archive-keyring.gpg\n"
-                    )
-            except IOError:
-                self.log.exception("failed to open %s" % sources_file)
-                raise FllError
-            finally:
-                if sources_file_fh:
-                    sources_file_fh.close()
-
-    def prime_apt(self, chroot: str) -> None:
-        """Prepare apt for work in each build chroot. Fetch all required gpg
-        keys and initialize apt_pkg config."""
-        self.log.info(f"{chroot} - preparing apt...")
-        chroot_dir = os.path.join(self.temp, chroot)
-
-        apt_preferences = self.conf["options"].get("apt_preferences")
-        if apt_preferences:
-            self.log.info(
-                f"{chroot} - importing apt preferences file: {apt_preferences}"
-            )
-            try:
-                shutil.copy(apt_preferences, os.path.join(chroot_dir, "etc/apt/"))
-            except IOError:
-                self.log.error(f"failed to import apt preferences: {apt_preferences}")
-                raise FllError
-
-        sources_list = os.path.join(chroot_dir, "etc/apt/sources.list")
-        if os.path.isfile(sources_list):
-            os.unlink(sources_list)
-
-        src_uri = not self.opts.binary
-        self.write_apt_lists(chroot, cached=True, src_uri=src_uri)
-
-        keyrings = list()
-        for dist_repo in self.conf["chroots"][chroot]["repos"].keys():
-            repo = self.conf["chroots"][chroot]["repos"][dist_repo]
-            keyring = repo.get("keyring")
-            if keyring:
-                keyrings.append(keyring)
-
-        if keyrings:
-            self.apt_get(chroot, "update", insecure=True)
-            self.apt_get(chroot, "install", args=keyrings, insecure=True)
-
-        self.apt_get(chroot, "update")
-        self.apt_get(chroot, "dist-upgrade")
-
-    def dpkg_divert(self, chroot: str) -> None:
-        """Divert some facilities and replace temporaily with /bin/true (or
-        some other more appropiate facility."""
-        chroot_dir = os.path.join(self.temp, chroot)
-        for divert in self.diverts:
-            self.log.debug(f"diverting {divert}")
-            self.chroot_exec(chroot, [
-                "dpkg-divert", "--add", "--local",
-                "--divert", f"{divert}.REAL", "--rename", divert,
-            ])
-
-            if divert == "/usr/sbin/policy-rc.d":
-                self.write_file(chroot_dir, divert)
-                os.chmod(os.path.join(chroot_dir, divert.lstrip("/")), 0o755)
-            else:
-                os.symlink("/bin/true", os.path.join(chroot_dir, divert.lstrip("/")))
-
-    def dpkg_undo_divert(self, chroot: str) -> None:
-        """Undivert facilities diverted by self.dpkg_divert()."""
-        chroot_dir = os.path.join(self.temp, chroot)
-        for divert in self.diverts:
-            self.log.debug(f"undoing diversion: {divert}")
-            os.unlink(os.path.join(chroot_dir, divert.lstrip("/")))
-            self.chroot_exec(chroot, ["dpkg-divert", "--remove", "--rename", divert])
 
     def write_file(self, chroot: str, filename: str, mode: int = 0o644) -> None:
         """Write a file in a chroot. Templates defined below."""
@@ -951,372 +392,24 @@ class FLLBuilder(BootloaderMixin):
                 self.log.exception("homed configuration failed")
                 raise FllError
 
-    def create_initramfs(self, chroot: str) -> None:
-        """Create an initramfs"""
-        initramfs_tool = self.conf["options"].get("initramfs_tool")
-        kvers = self.detect_linux_version(chroot)
-        for kernel in kvers:
-            cmd = ""
-            if initramfs_tool == "initramfs-tools":
-                cmd = ["update-initramfs", "-c", "-k", kernel]
-                if self.opts.verbose or self.opts.debug:
-                    cmd.append("-v")
-            elif initramfs_tool == "dracut":
-                cmd = [
-                    "dracut",
-                    "--no-hostonly",
-                    "--no-hostonly-i18n",
-                    "--no-hostonly-cmdline",
-                    "--no-hostonly-default-device",
-                    "--force-add", "fll",
-                    "--kver", kernel,
-                ]
-                if self.opts.verbose or self.opts.debug:
-                    cmd.append("--verbose")
-                elif self.opts.quiet:
-                    cmd.append("--quiet")
-            self.chroot_exec(chroot, cmd)
+    def hashsum(self, filename: str) -> str:
+        """Return SHA-256 hex digest of a file."""
+        h = hashlib.sha256()
+        with open(filename, "rb") as f:
+            for chunk in iter(lambda: f.read(65536), b""):
+                h.update(chunk)
+        return h.hexdigest()
 
-    def preseed_debconf(self, chroot: str) -> None:
-        """Preseed debconf with values read from package lists."""
-        chroot_dir = os.path.join(self.temp, chroot)
-        debconf_list = self.profiles[chroot].debconf
-
-        if not debconf_list:
-            return
-
-        self.log.debug(f"{chroot} - preseeding debconf...")
-        debconf_filename = os.path.join(chroot_dir, "fll", "fll_debconf_selections")
-        with open(debconf_filename, "w") as debconf_fh:
-            debconf_fh.writelines([f"{d}\n" for d in debconf_list])
-
-        cmd = ["debconf-set-selections"]
-        if self.opts.verbose:
-            cmd.append("--verbose")
-        cmd.append("/fll/fll_debconf_selections")
-        self.chroot_exec(chroot, cmd)
-
-    def detect_linux_version(self, chroot: str) -> list:
-        """Return version string of a singularly installed linux-image."""
-        linux = self.conf["chroots"][chroot]["packages"]["linux"]
-        linux_meta = f"linux-image-{linux}"
-        linux_images = [
-            f[len("linux-image-") :]
-            for f in self.profiles[chroot].manifest
-            if f.startswith("linux-image-") and not f == linux_meta
-        ]
-
-        if len(linux_images) > 0:
-            linux_images.sort(key=Version, reverse=True)
-            return linux_images
-
-        self.log.critical(f"{chroot} - failed to detect linux version")
-        raise FllError
-
-    def detect_locale_packages(
-        self, locales: list, wanted: list, available: dict
-    ) -> list:
-        """Provide automated detection for extra locales packages."""
-        self.log.debug(f"detecting packages for locales: {' '.join(locales)}")
-
-        locales_pkg_map_file = os.path.join(self.opts.share, "data", "locales-pkg-map")
-        locales_pkg_map = ConfigObj(locales_pkg_map_file)
-        self.log.debug("locales_pkg_map:")
-        self.log.debug(locales_pkg_map)
-
-        fll_locales = FllLocales(available, wanted, locales_pkg_map)
-        locales_list = []
-        for locale in sorted(locales):
-            try:
-                loc_pkg_list = fll_locales.detect_locale_packages(locale)
-            except FllLocalesError:
-                self.log.exception(f"Failed to parse locale string: {locale}")
-                raise FllError
-            else:
-                locales_list.extend(loc_pkg_list)
-
-        self.log.debug(f"locales_list: {' '.join(locales_list)}")
-        return locales_list
-
-    def detect_recommended_packages(
-        self, wanted: dict, available: dict, installed: set
-    ) -> list:
-        """Provide automated detection for packages in recommends whitelist."""
-        apt_recommends = self.conf["options"].get("apt_recommends")
-        if apt_recommends == "yes":
-            return []
-
-        self.log.debug("detecting whitelisted recommended packages...")
-        rec_module = ConfigObj(os.path.join(self.opts.share, "modules", "recommends"))
-        try:
-            rec_dict = dict(
-                [(p, True) for p in multiline_to_list(rec_module["packages"])]
-            )
-        except KeyError:
-            return []
-
-        self.log.debug(f"rec_dict: {rec_dict}")
-
-        rec_list = []
-        for p in wanted.keys():
-            if p not in rec_dict:
-                continue
-            pkg_data = available.get(p)
-            if not pkg_data:
-                continue
-            recommends_str = pkg_data.get("recommends", "")
-            if not recommends_str:
-                continue
-            for group in parse_dependency_groups(recommends_str):
-                if any(alt in wanted for alt in group):
-                    continue
-                first = group[0]
-                if first not in installed:
-                    self.log.debug(f"recommended package detected: {first} (via {p})")
-                    rec_list.append(first)
-        return rec_list
-
-    def _read_dpkg_status(self, chroot: str) -> dict:
-        """Parse /var/lib/dpkg/status; return {name: {version, source, source_version}}
-        for each installed package."""
-        status_file = os.path.join(self.temp, chroot, "var/lib/dpkg/status")
-        packages = {}
-        with open(status_file) as f:
-            for stanza_text in f.read().split("\n\n"):
-                stanza = {}
-                for line in stanza_text.splitlines():
-                    if line and not line.startswith(" "):
-                        key, _, val = line.partition(": ")
-                        stanza[key] = val
-                if "installed" not in stanza.get("Status", ""):
-                    continue
-                name = stanza.get("Package")
-                ver = stanza.get("Version")
-                if not name or not ver:
-                    continue
-                src = stanza.get("Source", name)
-                if "(" in src:
-                    srcver = src[src.index("(") + 1 : src.index(")")]
-                    src = src[: src.index(" (")]
-                else:
-                    srcver = ver
-                packages[name] = {
-                    "version": ver,
-                    "source": src,
-                    "source_version": srcver,
-                }
-        return packages
-
-    def _read_apt_packages(self, chroot: str) -> dict:
-        """Parse apt Packages files; return {name: {version, recommends}}
-        for all available packages, keeping the highest version when a package
-        appears in multiple index files."""
-        packages = {}
-        lists_dir = os.path.join(self.temp, chroot, "var/lib/apt/lists")
-        if not os.path.isdir(lists_dir):
-            return packages
-        for fname in sorted(os.listdir(lists_dir)):
-            if not fname.endswith("_Packages"):
-                continue
-            with open(os.path.join(lists_dir, fname)) as f:
-                for stanza_text in f.read().split("\n\n"):
-                    stanza = {}
-                    for line in stanza_text.splitlines():
-                        if line and not line.startswith(" "):
-                            key, _, val = line.partition(": ")
-                            stanza[key] = val
-                    name = stanza.get("Package")
-                    ver = stanza.get("Version")
-                    if not name or not ver:
-                        continue
-                    if name in packages:
-                        if Version(ver) <= Version(packages[name]["version"]):
-                            continue
-                    packages[name] = {
-                        "version": ver,
-                        "recommends": stanza.get("Recommends", ""),
-                    }
-        return packages
-
-    def write_manifest(self, chroot: str) -> None:
-        """Collect package and source package URI information from each
-        chroot."""
-        image_dir = os.path.join(
-            self.temp, "staging", self.conf["distro"]["FLL_IMAGE_DIR"]
-        )
-        self.log.info(f"{chroot} - writing package manifest...")
-
-        status = self._read_dpkg_status(chroot)
-        manifest = {
-            name: data["version"]
-            for name, data in status.items()
-            if not name.startswith("cdebootstrap-helper")
-        }
-        self.profiles[chroot].manifest = manifest
-
-        packages = list(manifest.keys())
-        packages.sort(key=len)
-        pkg_maxlen = len(packages[-1])
-        packages.sort()
-
-        manifest_name = self.get_distro_imagefile(chroot) + ".manifest"
-        manifest_file = os.path.join(image_dir, manifest_name)
-
-        try:
-            with open(manifest_file, "w") as manifest_fh:
-                manifest_fh.writelines(
-                    [
-                        f"{pkg.ljust(pkg_maxlen)} " + f"{manifest[pkg]}\n"
-                        for pkg in packages
-                    ]
-                )
-        except IOError:
-            self.log.exception(f"failed to write file: {manifest_file}")
-            raise FllError
-        finally:
-            os.chown(manifest_file, self.opts.uid, self.opts.gid)
-
-        if self.opts.binary:
-            return
-
-        self.log.info(f"{chroot} - writing source package URIs...")
-        srcpkg_seen = dict()
-        srcpkg_specs = []
-        for pkg in packages:
-            if pkg.startswith("cdebootstrap-helper"):
-                continue
-            srcpkg = status[pkg]["source"]
-            pkgver = status[pkg]["version"]
-            if srcpkg not in srcpkg_seen:
-                srcpkg_seen[srcpkg] = True
-                srcpkg_specs.append(f"{pkg}={pkgver}")
-
-        try:
-            output = self.chroot_output(
-                chroot, ["apt-get", "source", "--print-uris"] + srcpkg_specs
-            )
-        except FllError:
-            self.log.critical("failed to retrieve source URIs")
-            raise
-        uris = [
-            line.split("'")[1]
-            for line in output.splitlines()
-            if line.startswith("'")
-        ]
-        uris.sort()
-
-        sources_list = deduplicate_list(uris)
-
-        sources_name = self.get_distro_imagefile(chroot) + ".sources"
-        sources_file = os.path.join(image_dir, sources_name)
-
-        try:
-            with open(sources_file, "w") as sources_fh:
-                sources_fh.writelines([f"{s}\n" for s in sources_list])
-        except IOError:
-            self.log.exception(f"failed to write filename: {sources_file}")
-            raise FllError
-        finally:
-            os.chown(sources_file, self.opts.uid, self.opts.gid)
-
-    def install_packages(self, chroot: str) -> None:
-        """Install packages."""
-        available = self._read_apt_packages(chroot)
-        installed = set(self._read_dpkg_status(chroot).keys())
-
-        pkgs_base = list(installed)
-        pkgs_want = deduplicate_list(pkgs_base + list(self.profiles[chroot].packages))
-        pkgs_dict = dict([(pkg, True) for pkg in pkgs_want])
-        rec_pkgs = self.detect_recommended_packages(pkgs_dict, available, installed)
-        pkgs_want = deduplicate_list(list(pkgs_dict.keys()) + rec_pkgs)
-        pkgs_dict = dict([(pkg, True) for pkg in pkgs_want])
-        loc_list = (
-            [loc for loc in self.conf["chroots"][chroot]["packages"]["locales"]]
-            if self.conf["chroots"][chroot]["packages"].get("locales")
-            else self.opts.locales
-        )
-        loc_pkgs = self.detect_locale_packages(loc_list, pkgs_dict, available)
-        pkgs_want = deduplicate_list(list(pkgs_dict.keys()) + loc_pkgs)
-
-        self.log.info(f"{chroot} - installing packages...")
-        self.apt_get(chroot, "install", args=pkgs_want)
-
-    def install_flatpaks(self, chroot: str) -> None:
-        """Install flatpaks"""
-        if not any(
-            [self.profiles[chroot].flatpaks, self.profiles[chroot].flatpaks_beta]
-        ):
-            return
-        self.log.info(
-            f"{chroot} - configuring flatpak remotes and installing flatpaks..."
-        )
-        flatpak_base = ["flatpak"]
-        if self.opts.verbose:
-            flatpak_base.append("--verbose")
-        elif self.opts.debug:
-            flatpak_base.append("-vv")
-        flatpaks = self.profiles[chroot].flatpaks
-        if "flatpak" in self.profiles[chroot].packages:
-            self.chroot_exec(chroot, [
-                *flatpak_base, "remote-add", "--if-not-exists",
-                "flathub", "https://flathub.org/repo/flathub.flatpakrepo",
-            ])
-            for flatpak in flatpaks:
-                self.chroot_exec(chroot, [
-                    *flatpak_base, "install", "--noninteractive", "--assumeyes",
-                    "flathub", flatpak,
-                ])
-
-        flatpaks_beta = self.profiles[chroot].flatpaks_beta
-        if len(flatpaks_beta) > 0:
-            self.chroot_exec(chroot, [
-                *flatpak_base, "remote-add", "--if-not-exists",
-                "flathub-beta", "https://flathub.org/beta-repo/flathub-beta.flatpakrepo",
-            ])
-            for flatpak in flatpaks_beta:
-                self.chroot_exec(chroot, [
-                    *flatpak_base, "install", "--noninteractive", "--assumeyes",
-                    "flathub-beta", flatpak,
-                ])
-
-    def post_installation(self, chroot: str) -> None:
-        """Run package module postinst scripts in a chroot."""
-        chroot_dir = os.path.join(self.temp, chroot)
-
-        self.log.info(f"{chroot} - executing postinst scripts...")
-
-        for script in self.profiles[chroot].postinst:
-            sname = os.path.basename(script)
-            try:
-                shutil.copy(script, os.path.join(chroot_dir, "fll"))
-                os.chmod(os.path.join(chroot_dir, "fll", sname), 0o755)
-            except OSError:
-                self.log.exception(f"error preparing postinst script: {sname}")
-                raise FllError
-
-            self.chroot_exec(chroot, [f"/fll/{sname}", "postinst"])
-            os.unlink(os.path.join(chroot_dir, "fll", sname))
-
-    def zero_logs(self, chroot: str, dirname: str, filenames: list) -> None:
-        """Truncate all log files."""
-        chrootdir = dirname.partition(chroot)[2]
-
-        for f in filenames:
-            if not os.path.isfile(os.path.join(dirname, f)):
-                continue
-            self.write_file(chroot, os.path.join(chrootdir, f))
-
-    def clean_chroot(self, chroot: str) -> None:
-        """Remove unwanted content from a chroot."""
-        self.log.debug(f"{chroot} - purging unwanted content...")
-        chroot_dir = os.path.join(self.temp, chroot)
-
-        self.chroot_exec(chroot, ["dpkg", "--purge", "fll-live-initramfs"])
-        self.chroot_exec(chroot, ["apt-get", "clean"])
-        self.chroot_exec(chroot, ["dpkg", "--clear-avail"])
-
-        for dirpath, dirnames, files in os.walk(os.path.join(chroot_dir, "var/log")):
-            self.zero_logs(chroot, dirpath, files)
+    def sign_file(self, filename) -> None:
+        """Sign a file with hashkey if available."""
+        if self.opts.hashkey:
+            self.log.info(f"signing file: {filename}")
+            cmd = ["gpg", "-s", "--default-key"]
+            cmd.append(self.opts.hashkey)
+            cmd.append(filename)
+            self.exec_cmd(cmd)
+        else:
+            self.log.info(f"not signing file (no key given): {filename}")
 
     def mkreadonlyfs_chroot(self, chroot: str) -> None:
         """Make readonly filesystem image of chroot."""
@@ -1401,25 +494,6 @@ class FLLBuilder(BootloaderMixin):
         except IOError:
             self.log.exception("failed to move readonly rootfs image to staging dir")
             raise FllError
-
-    def hashsum(self, filename: str) -> str:
-        """Return SHA-256 hex digest of a file."""
-        h = hashlib.sha256()
-        with open(filename, "rb") as f:
-            for chunk in iter(lambda: f.read(65536), b""):
-                h.update(chunk)
-        return h.hexdigest()
-
-    def sign_file(self, filename) -> None:
-        """Sign a file with hashkey if available."""
-        if self.opts.hashkey:
-            self.log.info(f"signing file: {filename}")
-            cmd = ["gpg", "-s", "--default-key"]
-            cmd.append(self.opts.hashkey)
-            cmd.append(filename)
-            self.exec_cmd(cmd)
-        else:
-            self.log.info(f"not signing file (no key given): {filename}")
 
     def gen_live_media(self) -> None:
         """Generate live media iso image."""
